carts: Simulation-Based Assessment of Covariate Adjustment in Randomized Trials

Monte Carlo simulation framework for different randomized clinical trial designs with a special emphasis on estimators based on covariate adjustment.

Monte Carlo simulation framework for different randomized clinical trial designs with a special emphasis on estimators based on covariate adjustment.

See also

Author

Maintainer: Benedikt Sommer benediktsommer92@gmail.com

Authors:

Klaus K. Holst klaus@holst.it

Foroogh Shamsi foroogh.shamsi1@gmail.com

Benedikt Sommer, Klaus Holst, Foroogh Shamsi

Examples

if (FALSE) { # \dontrun{
trial <- Trial$new(
  covariates = \(n) data.frame(a = rbinom(n, 1, 0.5), x = rnorm(n)),
  outcome = setargs(outcome_count, par = c(1, 0.5, 1), overdispersion = 0.7)
)

trial$estimators(
  unadjusted = est_glm(family = "poisson"),
  adjusted = est_glm(family = "poisson", covariates = "x")
)

trial$run(n = 200, R = 100)
trial$summary()
} # }