Construct estimator for the treatment effect in RCT based on covariate adjustment

Efficient estimator of the treatment effect based on the efficient influence function. This involves a model for the conditional mean of the outcome variable given covariates.

Usage

est_adj(
  response = "y",
  treatment = "a",
  covariates = NULL,
  offset = NULL,
  id = NULL,
  family = gaussian(),
  level = 0.95,
  treatment.effect = NULL,
  nfolds = 1,
  ...
)

Arguments

response

Response variable (character)

treatment

Treatment variable (character)

covariates

Optional covariates (character)

offset

Optional offset (character)

id

Optional subject id variable (character)

family

Exponential family (default gaussian)

level

Confidence interval level

treatment.effect

(optional) function describing treatment effect given mean potential outcomes. Default is the average treatment effect (difference in expected outcome, (x,y) -> (y-x)) unless the family is binomial or poisson, in which case the treatment effect is the difference in log expected outcomes (x,y) -> log(y/x).

nfolds

Number of folds for estimating the conditional average treatment effect with double machine learning.

...

Additional arguments to targeted::learner_glm

Value

function

Details

...

See also

Author

Klaus Kähler Holst