Construct estimator for the treatment effect in RCT

Returns an estimator for the estimation of a treatment effect with parametric models (stats::glm and MASS::glm.nb). The returned estimator is a function with a single argument (data) and returns the treatment effect estimate, which is estimated with lava::estimate

Usage

est_glm(
  response = "y",
  treatment = "a",
  covariates = NULL,
  offset = NULL,
  id = NULL,
  level = 0.95,
  family = gaussian(),
  target.parameter = treatment,
  ...
)

Arguments

response

(character) Response variable

treatment

(character) Treatment variable

covariates

(character; optional) Single or vector of covariates

offset

(character; optional) Model offset

id

(character; optional) Subject id variable

level

(numeric) Confidence interval level

family

(family or character) Exponential family that is supported by stats::glm and MASS::glm.nb

target.parameter

(character) Target parameter from model output

...

Additional arguments to lava::estimate

Value

function

See also

Author

Klaus Kähler Holst

Examples

trial <- Trial$new(
    covariates = function(n) data.frame(a = rbinom(n, 1, 0.5), x = rnorm(n)),
    outcome = setargs(outcome_count,
      mean = ~ 1 + a*x,
      par = c(1, -0.1, 0.5, 0.2),
      overdispersion = 2)
)
dd <- trial$simulate(3e2)

# crude mean comparison between arms (default behavior; y ~ a)
est <- est_glm(family = poisson)
est(dd)
#>   Estimate Std.Err    2.5%  97.5% P-value
#> a  -0.1584  0.1915 -0.5337 0.2169  0.4082

# linear adjustment with one covariate (y ~ a + x)
est <- est_glm(family = poisson, covariates = "x")
est(dd)
#>   Estimate Std.Err    2.5%  97.5% P-value
#> a  -0.2085  0.1848 -0.5706 0.1537  0.2593

# return estimates of all linear coefficients (useful for debugging)
est <- est_glm(family = poisson, covariates = "x", target.parameter = NULL)
est(dd)
#>             Estimate Std.Err    2.5%  97.5%   P-value
#> (Intercept)   1.3372 0.12877  1.0848 1.5896 2.909e-25
#> a            -0.2085 0.18478 -0.5706 0.1537 2.593e-01
#> x             0.4094 0.08995  0.2331 0.5857 5.338e-06

# comparing robust and non-robust standard errors of poisson estimator by
# passing robust argument via ... to lava::estimate
estimators <- list(
  robust = est_glm(family = poisson),
  non.robust = est_glm(family = poisson, robust = FALSE)
)
res <- do.call(rbind, lapply(estimators, \(est) est(dd)$coefmat))
rownames(res) <- names(estimators)
res
#>              Estimate    Std.Err       2.5%       97.5%     P-value
#> robust     -0.1583593 0.19148563 -0.5336642  0.21694563 0.408235025
#> non.robust -0.1583593 0.05994537 -0.2758501 -0.04086853 0.008248453