EXPERIMENTAL: Outcome model for recurrent events with terminal events end-points

This function is still in an experimental state where the interface and functionality might change in the future

Usage

outcome_recurrent(
  data,
  lp = NULL,
  par = NULL,
  outcome.name = c("time", "status"),
  remove = c("id", "num"),
  model = NULL,
  death.model = NULL,
  death.lp = NULL,
  death.par = NULL,
  cens.model = NULL,
  cens.lp = NULL,
  cens.par = NULL,
  ...
)

Arguments

data

data.frame (covariates)

lp

linear predictor (formula or function)

par

optional list of model parameter

outcome.name

names of outcome (time and censoring status)

remove

variables that will be removed from input data (if formula is not specified)

model

optional mets::phreg object

death.model

optional model for death (terminal) events

death.lp

optional death linear predictor argument (formula or function)

death.par

optional list of death model parameters

cens.model

optional model for censoring mechanism

cens.lp

optional censoring linear predictor argument (formula or function)

cens.par

optional list of censoring model parameters

...

Additional arguments to outcome_lp

Value

function (random generator)

See also