Add additional covariates to existing list of covariates

For use with Trial objects, this function makes it possible to easily add additional covariates to an existing list of covariates (in the form of a data.frame or data.table).

Usage

covar_add(covars, x, names = NULL, ...)

Arguments

covars

list of covariates (data.frame's or data.table's)

x

new covariates (function or list of functions/scalars)

names

optional names of new covariates

...

additional arguments to function x or functions in x

Value

matching format of covariates in covars

Author

Klaus Kähler Holst

Examples

# adding "fixed" treatment indicator in each period
n <- 5
xt <- function(n, ...) {
 covar_loggamma(n, normal.cor = 0.2) |>
   covar_add(list(a = 0, a = 1))
}
xt(n)
#> $`0`
#>               z     a
#>           <num> <num>
#> 1: -1.314957908     0
#> 2:  0.437148133     0
#> 3:  0.886258819     0
#> 4:  0.008682371     0
#> 5:  0.016494232     0
#> 
#> $`1`
#>              z     a
#>          <num> <num>
#> 1: -0.14046420     1
#> 2:  0.48415947     1
#> 3: -3.25286323     1
#> 4: -0.11741507     1
#> 5:  0.01759809     1
#> 
# adding randomized treatment indicator
xt <- function(n, ...) {
 covar_loggamma(n, normal.cor = 0.2) |>
   covar_add(list(a = rbinom(n, 1, 0.5), a = rbinom(n, 1, 0.5)))
}
xt(5)
#> $`0`
#>              z     a
#>          <num> <int>
#> 1:  0.27939160     0
#> 2:  1.23280250     1
#> 3: -2.81056202     1
#> 4:  1.07608473     1
#> 5:  0.07663859     0
#> 
#> $`1`
#>             z     a
#>         <num> <int>
#> 1:  0.1135722     1
#> 2:  1.6015547     0
#> 3: -0.5877852     1
#> 4: -0.4335034     1
#> 5:  1.1675433     0
#> 
# adding baseline covariates
xt <- function(n, ...) {
 covar_loggamma(n, normal.cor = 0.2) |>
   covar_add(rnorm(n), names = "w1") |> # data
   covar_add(list(w2 = rnorm(n))) |> # data
   covar_add(data.frame(w3 = rnorm(n))) |> # data
   covar_add(\(n) data.frame(w4 = rnorm(n))) |> # function
   covar_add(\(n) rnorm(n), names = "w5") # function
}
xt(5)
#> $`0`
#>             z         w1         w2         w3          w4           w5
#>         <num>      <num>      <num>      <num>       <num>        <num>
#> 1: -1.1597797  0.1384708  1.2362631 -0.1609534 -0.54531347  0.403167147
#> 2:  0.1252903 -1.8252874  0.8806714 -0.1733009  0.08089318  0.008137872
#> 3: -2.5187935  1.1037974  0.9313603 -0.2917484 -0.36807541 -0.356770804
#> 4: -0.4206908  0.7713715  1.3684064  2.2466688  0.45485114  0.494554007
#> 5:  1.4236045  0.9091486 -1.5394573  0.3710951 -0.27716448  1.035993007
#> 
#> $`1`
#>             z         w1         w2         w3          w4           w5
#>         <num>      <num>      <num>      <num>       <num>        <num>
#> 1: -0.3529190  0.1384708  1.2362631 -0.1609534 -0.54531347  0.403167147
#> 2: -1.4938520 -1.8252874  0.8806714 -0.1733009  0.08089318  0.008137872
#> 3: -0.8087580  1.1037974  0.9313603 -0.2917484 -0.36807541 -0.356770804
#> 4:  0.4098094  0.7713715  1.3684064  2.2466688  0.45485114  0.494554007
#> 5:  0.9904463  0.9091486 -1.5394573  0.3710951 -0.27716448  1.035993007
#>