Add additional covariates to existing covariate random generator

For use with Trial objects, this function makes it possible to easily add additional covariates to an existing random generator (function(n ...) returning a data.frame or data.table)

Usage

covar_join(f, ...)

Arguments

f

covariate random generator

...

additional covariate generators or constant covariates

Value

function, with returned data type matching that of f

Examples

# single period
n <- 5
c1 <- function(n) data.frame(a = rnorm(n))
c2 <- function(n) data.frame(b = rnorm(n))
x <- c1 %join% c2
x(n)
#>            a          b
#> 1 -1.2683205  0.8395000
#> 2  0.6346674 -0.3503503
#> 3  0.2799669  1.1048359
#> 4  1.1000911 -0.2764584
#> 5  0.6364982  0.1311770

# adding covariates that remain constant when sampling
x <- c1 %join% data.frame(b = rnorm(n))
all.equal(x(n)$b, x(n)$b)
#> [1] TRUE

# adding multiple anonymous functions require parenthesis enclosing, with
# the exception of the last function
x <- c1 %join%
 (\(n) data.frame(b = rnorm(n))) %join%
 \(n) data.frame(c = rnorm(n))
x(n)
#>            a          b           c
#> 1 -0.8876138  0.3043006 -0.06007258
#> 2  1.6670633 -0.7345566  0.07290469
#> 3 -0.5793777 -0.5750394 -1.09248819
#> 4 -1.7808003  0.7322161 -1.12276716
#> 5  0.5400319 -0.7576544 -0.04935663

# multiple periods
base <- setargs(covar_loggamma, normal.cor = .5)
x <- base %join%
  function(n) list(
      data.frame(a = rbinom(n, 1, 0.5)),
      data.frame(a = rbinom(n, 1, 0.5))
    )
x(n)
#> $`0`
#>             z     a
#>         <num> <int>
#> 1:  0.3068251     1
#> 2:  0.5436833     0
#> 3:  0.1868201     0
#> 4:  0.8256138     0
#> 5: -4.0578567     0
#> 
#> $`1`
#>             z     a
#>         <num> <int>
#> 1: -0.1672489     0
#> 2: -0.8151929     0
#> 3: -0.7909171     1
#> 4: -0.4758204     0
#> 5: -1.6183318     0
#> 

# constant covariate
x <- base %join% list(data.frame(a = 0), data.frame(a = 1))
x(n)
#> $`0`
#>             z     a
#>         <num> <num>
#> 1: -4.9533081     0
#> 2: -0.4464893     0
#> 3:  0.2572230     0
#> 4:  1.4374533     0
#> 5: -1.3194759     0
#> 
#> $`1`
#>             z     a
#>         <num> <num>
#> 1:  0.2635741     1
#> 2: -0.8414074     1
#> 3: -0.1020544     1
#> 4:  0.5415538     1
#> 5:  0.7452187     1
#> 

# baseline covariate
x <- base %join% function(n) data.frame(w = rnorm(n))
x(n)
#> $`0`
#>             z          w
#>         <num>      <num>
#> 1: -0.5872233 -2.0679367
#> 2: -1.9406215  1.0798266
#> 3:  1.0066129  0.1134837
#> 4: -1.9531890  0.6274909
#> 5: -2.2391743  0.8225039
#> 
#> $`1`
#>              z          w
#>          <num>      <num>
#> 1: -0.06306952 -2.0679367
#> 2:  0.60974029  1.0798266
#> 3: -0.12596608  0.1134837
#> 4: -1.59946611  0.6274909
#> 5: -0.41859697  0.8225039
#>