Simulate from count model given covariates

Simulate from count model with intensity $$\lambda = \text{exposure-time}\exp(\text{par}^\top X)$$ where \(X\) is the design matrix specified by the formula

Usage

outcome_count(
  data,
  mean = NULL,
  par = NULL,
  outcome.name = "y",
  exposure = 1,
  remove = c("id", "num"),
  zero.inflation = NULL,
  overdispersion = NULL,
  ...
)

Arguments

data

covariate matrix

mean

formula specifying design from 'x' or a function that maps x to the mean value. If NULL all main-effects of the covariates will be used.

par

Regression coefficients (default zero). Can be given as a named list corresponding to the column names of model.matrix

outcome.name

Name of outcome variable ("y")

exposure

Exposure times. Either a scalar, vector or function.

remove

Variables that will be removed from input x (if formula is not specified)

zero.inflation

vector of probabilities or a function of the covariates 'x' including an extra column 'rate' with the rate parameter.

overdispersion

variance of gamma-frailty either given as a numeric vector or a function of the covariates 'x' with an extra column 'rate' holding the rate parameter 'rate'

...

Additional arguments passed to lower level functions

Value

data.table with added exposure column

See also

Examples

set.seed(24)
covariates <- function(n) data.frame(a = rbinom(n, 1, 0.5))
outcome <- setargs(
  outcome_count,
  mean = ~ 1 + a,
  par = log(c(2.5, 0.65)),
  overdispersion = 1 / 2,
  exposure = 2 # identical exposure time for all subjects
)
trial <- Trial$new(covariates = covariates, outcome = outcome)
trial$simulate(5)
#>       id     a   num     y exposure
#>    <num> <int> <num> <num>    <num>
#> 1:     1     0     0     4        2
#> 2:     2     0     0     6        2
#> 3:     3     1     0     0        2
#> 4:     4     1     0     1        2
#> 5:     5     1     0     2        2
# alternating exposure times between subjects
trial$args_model(exposure = c(1, 2))
trial$simulate(5)
#>       id     a   num     y exposure
#>    <num> <int> <num> <num>    <num>
#> 1:     1     1     0     6        1
#> 2:     2     0     0     5        2
#> 3:     3     0     0     0        1
#> 4:     4     1     0     3        2
#> 5:     5     0     0     1        1
# treatment-dependent exposure times
trial$args_model(exposure = function(dd) 1 - 0.5 * dd$a)
trial$simulate(5)
#>       id     a   num     y exposure
#>    <num> <int> <num> <num>    <num>
#> 1:     1     0     0     2      1.0
#> 2:     2     0     0     6      1.0
#> 3:     3     0     0     1      1.0
#> 4:     4     0     0     1      1.0
#> 5:     5     1     0     1      0.5